Use syscalls directly if on Linux
