+++ src/platform/services.rs
@@ -41,11 +41,14 @@ pub fn install(service_functions_by_loader: usize) {
     unsafe {
         SERVICE_FUNCTIONS = service_functions_by_loader;
     }
-
 }
 #[inline(always)]
 unsafe fn addr(fn_id: usize) -> usize {
     core::ptr::read((SERVICE_FUNCTIONS + fn_id * core::mem::size_of::<usize>()) as *mut usize)
+}
+#[inline(always)]
+pub unsafe fn install_single_service(fn_id: usize, fn_ptr: usize) {
+    core::ptr::write((SERVICE_FUNCTIONS + fn_id * core::mem::size_of::<usize>()) as *mut usize, fn_ptr)
 }
 //#[inline(always)]
 pub unsafe fn alloc(size: usize, align: usize) -> *mut u8 {
@@ -71,7 +74,7 @@ pub unsafe fn realloc(ptr: *mut u8, old_size: usize, old_align: usize, new_size:
 pub fn exit(status: i32) -> ! {
     unsafe {
         let fn_ptr: native_func::D = core::mem::transmute(addr(5));
-        fn_ptr(status as usize)
+        fn_ptr(status as isize as usize)
     }
 }
 #[inline(always)]
